requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.0.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
envparse==0.2.0
idna==3.4
jsonschema==4.17.3
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.3
